+++ osu.Game.Rulesets.Bms/Beatmaps/BmsFolderImporter.cs
@@ -37,7 +37,7 @@ namespace osu.Game.Rulesets.Bms.Beatmaps
             this.storage = storage;
             songsStorage = storage.GetStorageForDirectory(SONGS_STORAGE_PATH);
             realmAccess = realm;
-            this.tableMd5Index = new Lazy<BmsTableMd5Index>(() => tableMd5Index ?? new BmsTableMd5Index(BmsDifficultyTableManager.GetShared(storage), realm));
+            this.tableMd5Index = new Lazy<BmsTableMd5Index>(() => tableMd5Index ?? new BmsTableMd5Index(BmsDifficultyTableManager.GetShared(storage)));
         }
 
         public Task<FolderImportResult> Import(ImportTask task, ImportParameters parameters = default, CancellationToken cancellationToken = default)
@@ -212,6 +212,7 @@ namespace osu.Game.Rulesets.Bms.Beatmaps
 
                 using var transaction = realm.BeginWrite();
                 managedExisting.DeletePending = false;
+                TableMd5Index.ApplyTo(managedExisting.Beatmaps);
                 transaction.Commit();
 
                 return managedExisting.ToLive(realmAccess);